BREAKING Returns HTTP code 413  when `post_max_size` is exceeded #11708
This case is not directly reported by PHP itself, neither via
`UploadedFileInterface`, so we try our best to detect the situation and
immediately returns the HTTP code 413. We do it immediately because when
the case happens, PHP will leave $_POST entirely empty so we have no way
to do anything further. So we let the client know via HTTP status and it
is up to him to handle in the best way possible.

When a 413 happens, the response body contains a detailed message
similar to:

```json
{"message": "The server `post_max_size` is configured to accept 10.00 MiB, but received 28.00 MiB"}
```

It is also BREAKING because `UploadMiddleware::processRequest()` is
dropped entirely without replacement. If you need to use this library
but you don't already use PSR-15 middleware, then you will have to
implement a workaround yourself.